# services/event-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Read side of the event catalogue: browse events and their seat maps.",
        "title": "event-service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/events": {
            "get": {
                "description": "Browse events newest first. Optionally restrict to upcoming events. Paginated with limit/offset.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "List events",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "only events that have not ended yet",
                        "name": "upcoming",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "rows to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.PaginatedEventsResponse"
                        }
                    },
                    "400": {
                        "description": "invalid limit/offset",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/{eventID}": {
            "get": {
                "description": "Fetch a single event by its UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get one event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "event UUID",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.EventResponse"
                        }
                    },
                    "400": {
                        "description": "eventID is not a UUID",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "no such event",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/{eventID}/seats": {
            "get": {
                "description": "The seat map for one event, ordered by section/row/number. Paginated with limit/offset.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "List an event's seats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "event UUID",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page size (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "rows to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.PaginatedSeatsResponse"
                        }
                    },
                    "400": {
                        "description": "eventID is not a UUID, or invalid limit/offset",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "no such event",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "http.EventResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "starts_at": {
                    "type": "string"
                },
                "venue": {
                    "type": "string"
                }
            }
        },
        "http.PaginatedEventsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.EventResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/http.PaginationMeta"
                }
            }
        },
        "http.PaginatedSeatsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.SeatResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/http.PaginationMeta"
                }
            }
        },
        "http.PaginationMeta": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "http.SeatResponse": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "price_minor": {
                    "type": "integer"
                },
                "row": {
                    "type": "string"
                },
                "section": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}